+++ src/CfdiUtils/Certificado/Certificado.php
@@ -4,6 +4,7 @@ namespace CfdiUtils\Certificado;
 
 use CfdiUtils\OpenSSL\OpenSSL;
 use CfdiUtils\OpenSSL\OpenSSLPropertyTrait;
+use CfdiUtils\Utils\RegimenCapitalRemover;
 
 class Certificado
 {
@@ -35,6 +36,9 @@ class Certificado
 
     /** @var string */
     private $pemContents;
+
+    /** @var string|null */
+    private $nameWithoutRegimenCapitalSuffix;
 
     /**
      * Certificado constructor.
@@ -80,6 +84,7 @@ class Certificado
         $this->certificateName = strval($data['name'] ?? '');
         $this->rfc = (string) strstr(($data['subject']['x500UniqueIdentifier'] ?? '') . ' ', ' ', true);
         $this->name = strval($data['subject']['name'] ?? '');
+        $this->nameWithoutRegimenCapitalSuffix = null;
         $serial = new SerialNumber('');
         if (isset($data['serialNumberHex'])) {
             $serial->loadHexadecimal($data['serialNumberHex']);
@@ -180,9 +185,22 @@ class Certificado
      * Name (Razón Social) set when certificate was created
      * @return string
      */
-    public function getName(): string
+    public function getName($trimSuffix = false): string
     {
-        return $this->name;
+        return (! $trimSuffix) ? $this->name : $this->getNameWithoutRegimenCapitalSuffix();
+    }
+
+    /**
+     * Name (Razón Social) set when certificate was created without *régimen de capital* suffix
+     */
+    public function getNameWithoutRegimenCapitalSuffix(): string
+    {
+        if (null === $this->nameWithoutRegimenCapitalSuffix) {
+            $remover = RegimenCapitalRemover::createDefault();
+            $this->nameWithoutRegimenCapitalSuffix = $remover->remove($this->name);
+        }
+
+        return $this->nameWithoutRegimenCapitalSuffix;
     }
 
     /**