+++ tests/CfdiUtilsTests/CfdiCreator40FromExistentTest.php
@@ -29,11 +29,11 @@ final class CfdiCreator40FromExistentTest extends TestCase
         $xmlSource = strval(file_get_contents($this->utilAsset('cfdi40-real.xml')));
         $nodeSource = XmlNodeUtils::nodeFromXmlString($xmlSource);
         $creator = CfdiCreator40::newUsingNode($nodeSource);
-        $creator->putCertificado(new Certificado($this->utilAsset('certs/EKU9003173C9.cer')), true);
+        $creator->putCertificado(new Certificado($this->utilAsset('certs/EKU9003173C9.cer')));
 
         $comprobante = $creator->comprobante();
         $this->assertCount(1, $comprobante->searchNodes('cfdi:Emisor'));
-        $this->assertSame('ESCUELA KEMPER URGATE SA DE CV', $comprobante->searchAttribute('cfdi:Emisor', 'Nombre'));
+        $this->assertSame('ESCUELA KEMPER URGATE', $comprobante->searchAttribute('cfdi:Emisor', 'Nombre'));
         $this->assertSame('EKU9003173C9', $comprobante->searchAttribute('cfdi:Emisor', 'Rfc'));
     }
 }